Restore binary plugins compatibility with fa2l (revert aced31fd)

+++ far/far2sdk/farplug-wide.h
@@ -588,7 +588,7 @@ struct OpenDlgPluginData
 
 struct DialogInfo
 {
-	size_t StructSize;
+	int StructSize;
 	GUID Id;
 	DWORD Owner;
 };
@@ -1759,7 +1759,7 @@ struct ViewerMode
 
 struct ViewerInfo
 {
-	size_t StructSize;
+	int    StructSize;
 	int    ViewerID;
 	const wchar_t *FileName;
 	int64_t FileSize;
@@ -2395,7 +2395,7 @@ enum BOX_DEF_SYMBOLS
 
 typedef struct FarStandardFunctions
 {
-	size_t StructSize;
+	int StructSize;
 
 	FARSTDATOI                 atoi;
 	FARSTDATOI64               atoi64;
@@ -2472,7 +2472,7 @@ typedef struct FarStandardFunctions
 
 struct PluginStartupInfo
 {
-	size_t StructSize;
+	int StructSize;
 	const wchar_t *ModuleName;
 	INT_PTR ModuleNumber;
 	const wchar_t *RootKey;
@@ -2539,7 +2539,7 @@ enum PLUGIN_FLAGS
 
 struct PluginInfo
 {
-	size_t StructSize;
+	int StructSize;
 	DWORD Flags;
 	const wchar_t * const *DiskMenuStrings;
 	int *Reserved0;
@@ -2653,7 +2653,7 @@ enum OPERATION_MODES
 
 struct OpenPluginInfo
 {
-	size_t                        StructSize;
+	int                           StructSize;
 	DWORD                         Flags;
 	const wchar_t*                HostFile;
 	const wchar_t*                CurDir;
